add auth with bcrypt and JWT

+++ handlers/notes.go
@@ -6,11 +6,14 @@ import (
 	"net/http"
 	"strconv"
 	"strings"
-
+	"golang.org/x/crypto/bcrypt"
 	"notes-api/db"
 	"notes-api/models"
+	"github.com/golang-jwt/jwt/v5"
+	"time"
+	"os"
 )
-
+var jwtKey = []byte(os.Getenv("JWT_SECRET"))
 func GetIDFromPath(path string) (int, error) {
 	parts := strings.Split(path, "/")
 
@@ -21,9 +24,70 @@ func GetIDFromPath(path string) (int, error) {
 	return strconv.Atoi(parts[2])
 }
 
+func LoginHandler(w http.ResponseWriter, r *http.Request) {
+
+	if r.Method != "POST" {
+		http.Error(w, "Method not allowed", 405)
+		return
+	}
+
+	var user struct {
+		Username string `json:"username"`
+		Password string `json:"password"`
+	}
+
+	err := json.NewDecoder(r.Body).Decode(&user)
+	if err != nil {
+		http.Error(w, "Invalid JSON", 400)
+		return
+	}
+
+	if user.Username == "" || user.Password == "" {
+		http.Error(w, "Missing fields", 400)
+		return
+	}
+
+	var id int
+	var dbPassword string
+
+	err = db.DB.QueryRow(
+		"SELECT id, password FROM users WHERE username = ?",
+		user.Username,
+	).Scan(&id, &dbPassword)
+
+	if err != nil {
+		http.Error(w, "User not found", 401)
+		return
+	}
+
+	e:=bcrypt.CompareHashAndPassword([]byte(dbPassword),[]byte(user.Password))
+	
+		if e!=nil {
+		http.Error(w, "Invalid password", 401)
+		return
+	}
+	token:=jwt.NewWithClaims(jwt.SigningMethodHS256,jwt.MapClaims{
+		"user_id":id,
+		"exp":time.Now().Add(time.Hour*24).Unix(),
+	})
+	tokenString,err:=token.SignedString(jwtKey)
+	if err!=nil {
+		http.Error(w,"Token error",500)
+		return
+	}
+	json.NewEncoder(w).Encode(map[string]string{
+		"token":tokenString,
+	})
+	
+}
 func NotesHandler(w http.ResponseWriter, r *http.Request) {
 
 	w.Header().Set("Content-Type", "application/json")
+	userID, err := GetUserFromToken(r)
+		if err != nil {
+			http.Error(w, "Unauthorized", 401)
+			return
+		}
 
 	if r.Method == "GET" {
 
@@ -31,8 +95,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 
 		if err != nil {
 			if r.URL.Path == "/notes/" || r.URL.Path == "/notes" {
-
-				rows, err := db.DB.Query("SELECT id, text FROM notes")
+				rows, err := db.DB.Query("SELECT id, text FROM notes WHERE user_id =?",userID)
 				if err != nil {
 					http.Error(w, "DB error", 500)
 					return
@@ -56,8 +119,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 		}
 
 		var n models.Note
-
-		err = db.DB.QueryRow("SELECT id, text FROM notes WHERE id = ?", id).
+		err = db.DB.QueryRow("SELECT id, text FROM notes WHERE id = ? AND user_id=?", id,userID).
 			Scan(&n.ID, &n.Text)
 
 		if err != nil {
@@ -78,8 +140,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Invalid JSON", 400)
 			return
 		}
-
-		result, err := db.DB.Exec("INSERT INTO notes(text) VALUES(?)", n.Text)
+		result, err := db.DB.Exec("INSERT INTO notes(text,user_id) VALUES(?,?)", n.Text,userID,)
 		if err != nil {
 			http.Error(w, "DB error", 500)
 			return
@@ -90,7 +151,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 
 		json.NewEncoder(w).Encode(n)
 		return
-	}
+		}
 
 	if r.Method == "DELETE" {
 
@@ -99,8 +160,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Invalid ID", 400)
 			return
 		}
-
-		_, err = db.DB.Exec("DELETE FROM notes WHERE id = ?", id)
+		_, err = db.DB.Exec("DELETE FROM notes WHERE id = ? AND user_id=?", id,userID)
 		if err != nil {
 			http.Error(w, "DB error", 500)
 			return
@@ -127,8 +187,7 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Invalid JSON", 400)
 			return
 		}
-
-		_, err = db.DB.Exec("UPDATE notes SET text = ? WHERE id = ?", updated.Text, id)
+		_, err = db.DB.Exec("UPDATE notes SET text = ? WHERE id = ? AND user_id=?", updated.Text, id,userID)
 		if err != nil {
 			http.Error(w, "DB error", 500)
 			return
@@ -141,4 +200,78 @@ func NotesHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	http.Error(w, "Method not allowed", 405)
+}
+
+
+func SignupHandler(w http.ResponseWriter, r *http.Request) {
+
+	if r.Method != "POST" {
+		http.Error(w, "Method not allowed", 405)
+		return
+	}
+
+	var user struct {
+		Username string `json:"username"`
+		Password string `json:"password"`
+	}
+
+	err := json.NewDecoder(r.Body).Decode(&user)
+	if err != nil {
+		http.Error(w, "Invalid JSON", 400)
+		return
+	}
+
+	if user.Username == "" || user.Password == "" {
+		http.Error(w, "Missing fields", 400)
+		return
+	}
+
+	hash, err := bcrypt.GenerateFromPassword([]byte(user.Password), 10)
+	if err != nil {
+		http.Error(w, "Error hashing password", 500)
+		return
+	}
+
+	_, err = db.DB.Exec(
+		"INSERT INTO users(username,password) VALUES(?,?)",
+		user.Username,
+		string(hash),
+	)
+
+	if err != nil {
+		if strings.Contains(err.Error(), "UNIQUE") {
+			http.Error(w, "Username already exists", 400)
+			return
+		}
+		http.Error(w, "DB error", 500)
+		return
+	}
+
+	json.NewEncoder(w).Encode(map[string]string{
+		"message": "User created",
+	})
+}
+
+func GetUserFromToken(r *http.Request)(int,error) {
+	authHeader:=r.Header.Get("Authorization")
+	if authHeader == "" {
+		return 0,fmt.Errorf("missing token")
+	}
+	parts:=strings.Split(authHeader," ")
+	if(len(parts)!=2){
+		return 0,fmt.Errorf("invalid Header")
+	}
+	tokenStr:=parts[1]
+	token,err:=jwt.Parse(tokenStr,func(token *jwt.Token) (interface{}, error){
+		return jwtKey,nil
+	})
+	if err !=nil || !token.Valid {
+		return 0,fmt.Errorf("invalid token")
+	}
+	claims,ok:=token.Claims.(jwt.MapClaims)
+	if !ok {
+		return 0,fmt.Errorf("Invalid claims")
+	}
+	userID:= int(claims["user_id"].(float64))
+	return userID,nil
 }
